style : 반응형 구현 및 파일구조 변경
