+++ src/containers/index.ts
@@ -2,3 +2,4 @@ export { EditorMainPage } from './Editor/EditorMainPage';
 export { IntroMainPage } from './intro/IntroMainPage';
 export { InformationPage } from './InformationPage/InformationPage';
 export { DailyPage } from './DailyPage/DailyPage';
+export { MoreViewPage } from './moreViewPage/MoreViewPage';
